image_captioning_segmentation/requirements.txt
streamlit
torch
torchvision
transformers
Pillow
opencv-python